Started moving canvas-display to own module
 - This should make the main_window more easier to modify
